requirements.txt
pinecone-client==4.1.2
langchain-openai==0.1.16
openai==1.34.0
langchain==0.2.7